Non-linear Regression
If the data shows a curvy trend, then linear regression will not produce very accurate results when compared to a non-linear regression because, as the name implies, linear regression presumes that the data is linear.

Let's learn about non linear regressions and apply an example on python. In this notebook, we fit a non-linear model to the data points corresponding to China's GDP from 1960 to 2014.